Fixed headers

+++ include/structural_metrics/statement.h
@@ -5,7 +5,7 @@
 
 typedef struct {
     int total_statements;
-    int avg_statement_length;
+    double avg_statement_length;
 } StatementMetrics;
 
 void analyze_statements(const FileContent* fileData, StatementMetrics* metrics);
